fix naiming and modulation
