+++ js/data.js
@@ -1,3 +1,5 @@
+import {getRandomInteger, getRandomFloatNumber, getRandomElement, getSomeArray} from './util.js'
+
 const TYPE_HOUSING = ['palace', 'flat', 'house', 'bungalow'];
 const FEATURES_HOUSING = ['wifi','dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
 const ROOMS = [1, 2, 3, 100];
@@ -46,4 +48,53 @@ const Coordinate = {
 };
 const CHARACTERS = 5;
 
-export {TYPE_HOUSING, FEATURES_HOUSING, ROOMS, TIME_TYPE, PHOTOS_GALLARY, TITLES, DESCRIPTIONS, Price, AvatarId, Guests, Coordinate, CHARACTERS};
+// Созданием объект для массива
+const getAd = function() {
+  const registry = getRandomElement(TIME_TYPE);
+  const avatarId = getRandomInteger(AvatarId.MIN, AvatarId.MAX);
+
+  // Описание автора
+  const author = {
+    avatar: 'img/avatars/user0' + avatarId + '.png',
+  };
+
+  // Местоположение в виде географических координат
+  const location = {
+    x: getRandomFloatNumber(Coordinate.X.MIN, Coordinate.X.MAX, CHARACTERS),
+    y: getRandomFloatNumber(Coordinate.Y.MIN, Coordinate.Y.MAX, CHARACTERS),
+  };
+
+  // Иинформация об объявлении
+  const offer = {
+    title: getRandomElement(TITLES),
+    description: getRandomElement(DESCRIPTIONS),
+    address: Object.values(location).join(', '),
+    price: getRandomInteger(Price.MIN, Price.MAX),
+    type: getRandomElement(TYPE_HOUSING),
+    rooms: getRandomElement(ROOMS),
+    guests: getRandomInteger(Guests.MIN, Guests.MAX),
+    checkin: registry,
+    checkout: registry,
+    features: getSomeArray(FEATURES_HOUSING),
+    photos: getSomeArray(PHOTOS_GALLARY),
+  };
+
+  return {author, offer, location};
+};
+
+// Генерируем массив из 10 объектов
+const getAds = function() {
+  let ads = [];
+  let numberAds = 10;
+
+  for (let i = 0; i < numberAds; i++) {
+    ads.push(getAd());
+  }
+
+  return ads;
+};
+
+const ads = getAds();
+
+export {ads};
+
