+++ js/main.js
@@ -1,1 +1,4 @@
-import './array.js';
+import {ads} from './data.js';
+
+// eslint-disable-next-line no-console
+console.log(ads)
